Application now has an Admin panel
