+++ Blog/Controllers/AuthController.cs
@@ -31,7 +31,7 @@ namespace Blog.Controllers
         {
             var result = await _signInManager.PasswordSignInAsync(vm.UserName, vm.Password, false, false);
             
-            return RedirectToAction("Index", "Home");
+            return RedirectToAction("Index", "Panel");
         }
 
         [HttpGet]
