+++ Blog/Controllers/HomeController.cs
@@ -32,38 +32,6 @@ namespace Blog.Controllers
             return View(post);
         }
 
-        [HttpGet]
-        public IActionResult Edit(int? id)
-        {
-            if (id == null)
-                return View(new Post());
-            else
-            {
-                var post = _repo.GetPost((int)id);
-                return View(post);
-            }
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Edit(Post post)
-        {
-            if (post.Id > 0)
-                _repo.UpdatePost(post);
-            else
-                _repo.AddPost(post);
-
-            if (await _repo.SaveChangesAsync())
-                return RedirectToAction("Index");
-            else
-                return View(post);
-        }
-
-        [HttpGet]
-        public async Task<IActionResult> Remove(int id)
-        {
-            _repo.RemovePost(id);
-            await _repo.SaveChangesAsync();
-            return RedirectToAction("Index");
-        }
+       
     }
 }
