+++ Blog/Controllers/PanelController.cs
@@ -1,0 +1,63 @@
+﻿using Blog.Data.Repository;
+using Blog.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Blog.Controllers
+{
+    [Authorize(Roles = "Admin")]
+    public class PanelController : Controller
+    {
+        private IRepository _repo;
+
+        public PanelController(IRepository repo)
+        {
+            _repo = repo;
+        }
+
+        public IActionResult Index()
+        {
+            var posts = _repo.GetAllPost();
+            return View(posts);
+        }
+
+        [HttpGet]
+        public IActionResult Edit(int? id)
+        {
+            if (id == null)
+                return View(new Post());
+            else
+            {
+                var post = _repo.GetPost((int)id);
+                return View(post);
+            }
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> Edit(Post post)
+        {
+            if (post.Id > 0)
+                _repo.UpdatePost(post);
+            else
+                _repo.AddPost(post);
+
+            if (await _repo.SaveChangesAsync())
+                return RedirectToAction("Index");
+            else
+                return View(post);
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> Remove(int id)
+        {
+            _repo.RemovePost(id);
+            await _repo.SaveChangesAsync();
+            return RedirectToAction("Index");
+        }
+    }
+}
