+++ Blog/Views/Home/Index.cshtml
@@ -15,14 +15,6 @@
                 <a asp-controller="Home" asp-action="Post" asp-route-id="@post.Id">
                     @post.Title
                 </a>
-                -
-                <a asp-controller="Home" asp-action="Edit" asp-route-id="@post.Id">
-                    Edit
-                </a>
-                -
-                <a asp-controller="Home" asp-action="Remove" asp-route-id="@post.Id">
-                    Remove
-                </a>
             </li>
         }
     </ul>
