+++ Blog/Views/Panel/Index.cshtml
@@ -1,0 +1,27 @@
+﻿@model IEnumerable<Post>
+
+
+@{ 
+
+}
+<h2>Admin Panel</h2>
+
+
+<div>
+    <ul>
+        @foreach (var post in Model)
+        {
+            <li>
+                @post.Title
+                -
+                <a asp-controller="Panel" asp-action="Edit" asp-route-id="@post.Id">
+                    Edit
+                </a>
+                -
+                <a asp-controller="Panel" asp-action="Remove" asp-route-id="@post.Id">
+                    Remove
+                </a>
+            </li>
+        }
+    </ul>
+</div>
