+++ Blog/Views/Shared/_Layout.cshtml
@@ -11,6 +11,8 @@
         <a asp-controller="Home" asp-action="Index">Home</a>
         <a asp-controller="Home" asp-action="Edit">Create Post</a>
 
+        <a asp-controller="Panel" asp-action="Index">Admin Panel</a>
+
         @if (User.Identity.IsAuthenticated)
         {
             <a asp-controller="Auth" asp-action="Logout">Logout</a>
